[Chore] Start using test plans (#1417)
* Create separate test target for the integration tests

* Rename WireSyncEngine-iOS-Tests test target to UnitTests

* Add tests plans

* Clean up directory structure

* Wait for all tasks to complete before we tear down the contexts

* Clean up objects in tear down which otherwise would interfere with other tests

* Fix typo

* Fix retain cycle in the MockAppLock causing user session to not be released in tests

* Reset old requests before recreating SessionManager which would otherwise hold a reference to an already released MOC
