+++ Tests/Source/Integration/ClientManagementTests.m
@@ -19,7 +19,7 @@
 
 @import WireTesting;
 
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 @interface FakeClientObserver : NSObject <ZMClientUpdateObserver>
 @property (nonatomic) NSArray *fetchedClients;
