+++ Tests/Source/Integration/ConversationsTests.m
@@ -25,7 +25,7 @@
 
 #import <WireSyncEngine/WireSyncEngine-Swift.h>
 #import "ConversationTestsBase.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 
 @interface ZMConversationList (ObjectIDs)
