+++ Tests/Source/Integration/FileTransferTests.m
@@ -26,7 +26,7 @@
 #import "MessagingTest.h"
 #import "ZMConversationTranscoder+Internal.h"
 #import <WireSyncEngine/WireSyncEngine-Swift.h>
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 #import "ConversationTestsBase.h"
 
 @interface FileTransferTests : ConversationTestsBase
