+++ Tests/Source/Integration/IntegrationTest.m
@@ -22,7 +22,7 @@
 @import UIKit;
 
 #import "IntegrationTest.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 #import <WireSyncEngine/WireSyncEngine.h>
 
 @implementation IntegrationTest
