+++ Tests/Source/Integration/LoginFlowTests.m
@@ -22,7 +22,7 @@
 
 #import "NSError+ZMUserSessionInternal.h"
 #import <WireSyncEngine/WireSyncEngine-Swift.h>
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 #import "ZMCredentials.h"
 #import <WireSyncEngine/ZMAuthenticationStatus.h>
