+++ Tests/Source/Integration/SendAndReceiveMessagesTests.m
@@ -22,7 +22,7 @@
 @import WireDataModel;
 
 #import "ConversationTestsBase.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 
 @interface TestConversationObserver : NSObject <ZMConversationObserver>
