+++ Tests/Source/Integration/UserProfileTests.m
@@ -18,7 +18,7 @@
 
 #import "NSError+ZMUserSession.h"
 #import "NSError+ZMUserSessionInternal.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 #import "ZMClientRegistrationStatus+Internal.h"
 #import "ZMCredentials.h"
