+++ Tests/Source/Synchronization/Transcoders/ZMConnectionTranscoderTest.m
@@ -23,7 +23,7 @@
 #import "ObjectTranscoderTests.h"
 #import "ZMConnectionTranscoder+Internal.h"
 #import "ZMSimpleListRequestPaginator.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 @interface ZMConnectionTranscoderTest : ObjectTranscoderTests
 
@@ -58,8 +58,13 @@
 
 - (void)tearDown
 {
+    [(id)self.syncStateDelegate stopMocking];
+    self.syncStateDelegate = nil;
+
     [self.mockClientRegistrationDelegate tearDown];
     self.mockClientRegistrationDelegate = nil;
+
+    self.mockSyncStatus = nil;
     self.sut = nil;
     [super tearDown];
 }
