+++ Tests/Source/Synchronization/Transcoders/ZMSelfTranscoderTests.m
@@ -24,7 +24,7 @@
 
 #import "ObjectTranscoderTests.h"
 #import "ZMSelfStrategy+Internal.h"
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 
 @interface ZMSelfStrategyTests : ObjectTranscoderTests
