+++ Tests/Source/Synchronization/ZMOperationLoopTests.m
@@ -25,7 +25,7 @@
 #import "MessagingTest.h"
 #import "ZMSyncStrategy.h"
 #import <WireSyncEngine/WireSyncEngine-Swift.h>
-#import "WireSyncEngine_iOS_Tests-Swift.h"
+#import "Tests-Swift.h"
 #import "MockModelObjectContextFactory.h"
 #import "ZMOperationLoop+Private.h"
 #import "ZMSyncStrategy+Internal.h"
